backend/requirements.txt
aio-pika==9.5.6
aiormq==6.8.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==4.0.1
boto3==1.42.28
botocore==1.42.28
certifi==2026.1.4
cffi==2.0.0
click==8.1.8
cryptography==46.0.3
Deprecated==1.3.1
ecdsa==0.19.1
exceptiongroup==1.3.1
fastapi==0.128.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jmespath==1.0.1
limits==4.2
multidict==6.7.0
packaging==24.2
pamqp==3.3.0
passlib==1.7.4
propcache==0.4.1
psycopg2-binary==2.9.11
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-magic==0.4.27
python-multipart==0.0.20
rsa==4.9.1
s3transfer==0.16.0
sentry-sdk==2.49.0
six==1.17.0
slowapi==0.1.9
SQLAlchemy==2.0.45
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==1.26.20
uvicorn[standard]==0.39.0
wrapt==2.0.1
yarl==1.22.0
pytz
psutil==5.9.8
pandas
openpyxl
